% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel1.R
\name{log_par}
\alias{log_par}
\title{logarithm with parallel}
\usage{
log_par(x)
}
\arguments{
\item{x}{a vector to calculate logarithm}
}
\value{
logarithm of each coordinate of x
}
\description{
logarithm with parallel
}
\examples{
log_par(1:100)
}
