% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel1.R
\name{log_seq}
\alias{log_seq}
\title{logarithm sequentiel}
\usage{
log_seq(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector of logarithm of x
}
\description{
logarithm sequentiel
}
\examples{
log_seq(1:100)
}
