% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{x : une matrice, a n colonnes (les observations) et p lignes}

\item{mean}{un vecteur de moyennes}

\item{varcovM}{une matrice de variance-covariance}

\item{Log}{un paramètre logique valant TRUE par défaut}
}
\value{
renvoie une liste contenant la matrice x ainsi qu’un vecteur des images
des points de x par la fonction de densite de la variable aleatoire de loi normale multivariée consideree.
}
\description{
densite d’une loi normale multivariee sur R^p en n points.
}
\examples{
mvnpdf(x=matrix(1.96),log=FALSE)
dnorm(1.96)

}
