% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfoptim.R
\name{mvnpdfoptim}
\alias{mvnpdfoptim}
\title{mvnpdfoptim}
\usage{
mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{numeric matrix.}

\item{mean}{vector containing each mean of x's columns.}

\item{varcovM}{variance covariance matrix of x.}

\item{Log}{logical; if TRUE return the logarithm of multivariate distribution.}
}
\value{
a list containing values of x matrix and values of multivariate distribution calculated
}
\description{
Optimised function of mvnpdf. Calculates the multivariate distribution of a matrix
}
