% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfsmart.R, R/mvnpdfsmart_par.R
\name{mvnpdfsmart}
\alias{mvnpdfsmart}
\alias{mvnpdfsmart_par}
\title{mvnpdfsmart}
\usage{
mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart_par(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{numeric matrix.}

\item{mean}{vector containing each mean of x's columns.}

\item{varcovM}{variance covariance matrix of x.}

\item{Log}{logical; if TRUE return the logarithm of multivariate distribution.}
}
\value{
a list containing values of x matrix and values of multivariate distribution calculated

a list containing values of x matrix and values of multivariate distribution calculated
}
\description{
improve function of mvnpdf. Calculates the multivariate distribution of a matrix
}
