% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvnpdfsmartC}
\alias{mvnpdfsmartC}
\title{mvpdfsmartC}
\usage{
mvnpdfsmartC(x, mean, varcovM, Log = TRUE)
}
\arguments{
\item{x}{matrix x}

\item{mean}{column vector of x's mean}

\item{varcovM}{matrix of x's variance-covariance}

\item{Log}{boolean. If true, return log values of x's multivariate normal distribution}
}
\description{
mvpdfsmartC
}
