% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_dmvnorm.R
\name{my_dmvnorm}
\alias{my_dmvnorm}
\title{Copy of mvtnorm::dmvnorm for progiling and reference}
\usage{
my_dmvnorm(
  x,
  mean = rep(0, p),
  sigma = diag(p),
  log = FALSE,
  checkSymmetry = TRUE
)
}
\arguments{
\item{x}{matrix}

\item{mean}{vector}

\item{sigma}{covariance matrix}

\item{log}{logical. Default is \code{FALSE}}

\item{checkSymmetry}{logical. Default is \code{TRUE}}
}
\description{
Copy of mvtnorm::dmvnorm for progiling and reference
}
