% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkgr-package.R
\docType{package}
\name{mypkgr-package}
\alias{mypkgr}
\alias{mypkgr-package}
\title{mypkgr: mypkgr}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ErwanLouveau/mypkgr}
  \item \url{https://erwanlouveau.github.io/mypkgr/}
  \item Report bugs at \url{https://github.com/ErwanLouveau/mypkgr/issues}
}

}
\author{
\strong{Maintainer}: Erwan Louveau \email{erwan.louveau@u-bordeaux.fr}

}
\keyword{internal}
