% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkgr-package.r
\docType{package}
\name{mypkgr}
\alias{mypkgr-package}
\alias{mypkgr}
\title{mypkgr.}
\description{
This the place to speak about your package.
}
\details{
What is in the package...

What are the most important functions...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/borishejblum/mypkgr}
  \item \url{https://borishejblum.github.io/mypkgr/}
  \item Report bugs at \url{https://github.com/borishejblum/mypkgr/issues}
}

}
\author{
\strong{Maintainer}: Boris Hejblum \email{boris.hejblum@u-bordeaux.fr}

Authors:
\itemize{
  \item Robin Genuer \email{robin.genuer@u-bordeaux.fr}
}

}
