% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mvnpdfsmart_par.R
\name{plot.mvnpdfsmart_par}
\alias{plot.mvnpdfsmart_par}
\title{Plot of the mvnpdfsmart_par function}
\usage{
\method{plot}{mvnpdfsmart_par}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mvnpdfsmart} resulting from a call of the
\code{mnvpdf()} function.}

\item{...}{graphical parameters passed to \code{plot()} function.}
}
\value{
Nothing is returned, only a plot is given.
}
\description{
Plot of the mvnpdfsmart_par function
}
\examples{
pdfvalues <- mvnpdf(x=matrix(seq(-3, 3, by = 0.1), nrow = 1), Log=FALSE)
plot(pdfvalues)
}
