% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfsmart.R,
%   R/mvnpdfoptim.R, R/mvnpdf_invC.R, R/RcppExports.R, R/mvnpdfsmart_par.R,
%   R/mvnpdfsmart_parpb.R, R/mvnpdfsmart_sleepy_par.R, R/mvnpdfoptim_par.R,
%   R/mvnpdfsmart_sleepy.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfsmart}
\alias{mvnpdfoptim}
\alias{mvnpdf_invC}
\alias{mvnpdfsmartC}
\alias{mvnpdfoptimC}
\alias{mvnpdfC_LeChat}
\alias{mvnpdfsmart_par}
\alias{mvnpdfsmart_parpb}
\alias{mvnpdfsmart_sleepy_par}
\alias{mvnpdfoptim_par}
\alias{mvnpdfoptim_parIter}
\alias{mvnpdfsmart_sleepy}
\title{Multivariate-Normal probability density function}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdf_invC(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmartC(x, mean, varcovM, Log = TRUE)

mvnpdfoptimC(x, mean, varcovM, Log = TRUE)

mvnpdfC_LeChat(x, mean, varcovM, Log = TRUE)

mvnpdfsmart_par(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart_parpb(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE,
  ncpus = 1
)

mvnpdfsmart_sleepy_par(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)

mvnpdfoptim_par(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim_parIter(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE,
  ncores = 1
)

mvnpdfsmart_sleepy(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)
}
\arguments{
\item{x}{a p x n data matrix with n the number of observations and
p the number of dimensions}

\item{mean}{mean vector}

\item{varcovM}{variance-covariance matrix}

\item{Log}{logical flag for returning the log of the probability density
function. Default is \code{TRUE}.}

\item{ncpus}{number of cores for parallelized pbapply}

\item{ncores}{number of parallel cores to split task over when using itertools}
}
\value{
a list containing the input matrix x and y the multivariate-Normal probability density function
computed at x
}
\description{
This is a concise description of what the function does.
}
\details{
This part gives more details on the function.

the implementation of mvnpdfoptimC based on the implementation
from Nino Hardt and Dicko Ahmadou
http://gallery.rcpp.org/articles/dmvnorm_arma/ (accessed in August 2014)
}
\examples{
mvnpdf(x = matrix(1.96), Log = FALSE)
dnorm(1.96)

mvnpdf(x = matrix(rep(1.96, 2), nrow = 2, ncol = 1), Log = FALSE)
\dontrun{
n <- 10000
mb <- microbenchmark::microbenchmark(
  mvtnorm::dmvnorm(matrix(1.96, nrow = n, ncol = 2)),
  mvnpdfsmart(x=matrix(1.96, nrow = 2, ncol = n), Log=FALSE),
  mvnpdfoptim_par(x=matrix(1.96, nrow = 2, ncol = n), Log=FALSE),
  times=10L)
mb
}

}
