% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{fonction normale multivariee pour la densite
 description}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{les points}

\item{mean}{les moyennes}

\item{varcovM}{la matrice de var cov sigma}

\item{Log}{si on veut les résultats en log, par défaut = TRUE}
}
\value{
les densites
}
\description{
fonction normale multivariee pour la densite
 description
}
\examples{
mvnpdf(x=matrix(1.96), Log=FALSE)
dnorm(1.96)
}
