% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfRAW.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix}

\item{mean}{a vector of means}

\item{varcovM}{a variance-covariance matrix}

\item{Log}{a logical parameter. Default is TRUE}
}
\value{
a list containing the matrix x, and a vector of length n of the multivariate normal distribution density values at those points
}
\description{
compute the value of the density of a multivariate normal distribution on Rp at n points
}
\examples{
mvnpdf(x=matrix(1.96), Log=FALSE)
dnorm(1.96)
mvnpdf(x=matrix(rep(1.96, 2), nrow=2, ncol=1), Log=FALSE)
}
