% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Fonction mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{: matrice d'observations en colonnes (n colonnes, p lignes)}

\item{mean}{: vecteur de moyenne}

\item{varcovM}{: matrice de variance covariance}

\item{Log}{: TRUE par défaut (calcul du log)}
}
\value{
matrice x et vecteur y de la valeur de la densité en chacun des vecteurs colonne de x
}
\description{
calcule la valeur de la densité d’une loi normale multivariée sur Rp en n points
}
