% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Fonction de densité multivariée}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{matrice de p lignes et n colonnes (n=observations)}

\item{mean}{vecteur de p moyennes}

\item{varcovM}{matrice carrée de variance covariance de dimension p}

\item{Log}{dans l'échelle logarithme (facilite le calcul)}
}
\value{
: une liste : points x et les valeurs de la densité évaluées sur ces points
}
\description{
Calcul de la densité d’une loi normale multivariée
}
\details{
This gives more details on the function
}
