% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Densité de loi normale multivariée}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{pxn matrice avec n le nombre de d'observation et p le nombre de dimension}

\item{mean}{vecteur de moyenne de la gaussienne (de taille px1)}

\item{varcovM}{matrice de covariance de la gaussienne (de taille pxp)}

\item{Log}{si True, retournera le logarithme des densités. Par défaut, la valeur est TRUE.}
}
\value{
une liste contenant la matrice d'entrée x et y les valeurs de la densité
de la gaussienne multivariée en x
}
\description{
Calcule les images par la fonction de densité d'une gaussienne multivariée
}
\examples{
mvnpdf(x=matrix(1.96), Log=FALSE)
dnorm(1.96)

mvnpdf(x=matrix(rep(1.96, 2), nrow=2, ncol=1), Log=FALSE)

}
