% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mvnpdf.R
\name{mvnpdfsmartC}
\alias{mvnpdfsmartC}
\alias{mvnpdf}
\title{MVNPDF}
\usage{
mvnpdfsmartC(x, mean, varcovM, Log = TRUE)

mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix, with n columns (the observations) and p rows.}

\item{mean}{a vector of means.}

\item{varcovM}{a variance-covariance matrix.}

\item{Log}{a logical parameter, with default value to TRUE.}
}
\value{
Returns a list containing the matrix x, and a vector of length n of the multivariate normal distribution density values at those points (eventually in log-scale).
}
\description{
Description de la fonction.
}
\details{
Détails de la fonction.
}
\examples{
mvnpdf(x = matrix(1.96), Log = FALSE)
dnorm(1.96)
}
