% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfoptim.R, R/mvnpdfsmart.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart}
\title{Densite d une loi normale multivariee}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)
}
\arguments{
\item{x}{matrice taille p*n (n observations et p dimensions)}

\item{mean}{vecteur de moyennes}

\item{varcovM}{une matrice de variance-covariance}

\item{Log}{booleen pour retourner le log}
}
\value{
renvoie une liste contenant la matrice x ainsi qu un vecteur
des images des points de x par la fonction de densite de la variable
aleatoire de loi normale multivariee consideree.
}
\description{
Densite d une loi normale multivariee
}
\examples{
mvnpdf(x=matrix(1.96), log=FALSE)
dnorm(1.96)

}
