% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfRAW.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Title}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{is a matrix containing the observations in columns}

\item{mean}{mean vector of the multivariate gaussian distribution}

\item{varcovM}{variance covariance matrix}

\item{Log}{logical parameter (default TRUE)}
}
\value{
list of the results
}
\description{
description
}
\details{
details
}
