% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{data matrix}

\item{mean}{mean vector of the multivariate normal distribution}

\item{varcovM}{variance-covariance matrix}

\item{Log}{if TRUE (default) returns the log scale}
}
\value{
a list of length 2 with the data matrix as the first argument
and a vector of length n with the multivariate normal distribution
density values at the points in x.
}
\description{
Compute the value of the density of a multivariate normal distribution
on R^p at n points.
}
\details{
Details
}
\examples{
example
}
