% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfRox.R, R/mvnpdfoptim.R,
%   R/mvnpdfsmart.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart}
\title{mvnpdf}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)
}
\arguments{
\item{x}{x : une matrice, a n colonnes (les observations) et p lignes}

\item{mean}{un vecteur de moyennes}

\item{varcovM}{une matrice de variance-covariance}

\item{Log}{un paramètre logique valant TRUE par défaut}

\item{x}{p x n data matrix with n the number of observations and
p the number of dimensions}

\item{mean}{mean vector or list of mean vectors (either a vector or
a matrix)}

\item{varcovM}{variance-covariance matrix or list of variance-covariance
matrices (either a matrix or a list)}

\item{Log}{logical flag for returning the log of the probability density
function. Default is \code{TRUE}.}
}
\value{
renvoie une liste contenant la matrice x ainsi qu’un vecteur des images
des points de x par la fonction de densite de la variable aleatoire de loi normale multivariée consideree.

a list containing the input matrix x and y the multivariate-Normal probability density function
computed at x
}
\description{
densite d’une loi normale multivariee sur R^p en n points.

This is a concise description of what the function does.
}
\details{
This part gives more details on the function.
}
\examples{
mvnpdf(x=matrix(1.96),log=FALSE)
dnorm(1.96)

mvnpdf(x=matrix(1.96), Log=FALSE)
dnorm(1.96)

mvnpdf(x=matrix(rep(1.96, 2), nrow=2, ncol=1), Log=FALSE)
}
