% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Multivariate Normal Probability Density Function}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{matrice des antecedents : obs en colonnes}

\item{mean}{vecteur esperance}

\item{varcovM}{matrice covariance}

\item{Log}{TRUE pour renvoyer log y}
}
\value{
renvoie x et y ou log y si Log TRUE
}
\description{
Compute multivariate gaussian probability distribution function
}
\details{
Details
}
