% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypkgr-package.R
\docType{package}
\name{mypkgr-package}
\alias{mypkgr}
\alias{mypkgr-package}
\title{mypkgr: Example Of A Package For Teaching The EndeavR Course}
\description{
A collection of function for computing the density for a multivariate Gaussian distibution. Several implementations are provided to examplify the different techniques for optimizing and accelerating R codes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/borishejblum/mypkgr}
  \item \url{https://borishejblum.github.io/mypkgr/}
  \item Report bugs at \url{https://github.com/borishejblum/mypkgr/issues}
}

}
\author{
\strong{Maintainer}: Boris Hejblum \email{boris.hejblum@u-bordeaux.fr}

Authors:
\itemize{
  \item Robin Genuer \email{robin.genuer@u-bordeaux.fr}
}

}
\keyword{internal}
